Add an option to save the Qt archive files
By directly running `./dk SonicScout_Setup.Qt windows_x86_64` the downloaded files are saved to an archive directory. The metadata for those files (where it can be downloaded, and a checksum) are saved in a JSON file. That JSON file is captured in the dk-archive-server project so there is a permanent mirror available at some URL (maybe https://diskuv.com/archives/qt/...) that replicates just the parts of http://mirrors.ocf.berkeley.edu/qt/ that are needed.

+ Remove duplicate uv arguments
+ Ignore any Qt 5 or Qt 6 directory in git
+ Use correct ocamlformat disable directive